// src/types/role.ts
export enum Role {
  ADMIN = 'ADMIN',
  MANAGER = 'MANAGER',
  COLABORATOR = 'COLABORATOR',
  STUDENT = 'STUDENT',
  MEDICAL = 'MEDICAL',
  NURSE = 'NURSE',
  NURSING_TECHNICIAN = 'NURSING_TECHNICIAN',
  PSYCHOLOGIST = 'PSYCHOLOGIST',
  NUTRITIONIST = 'NUTRITIONIST',
  RESPONSABLE = 'RESPONSABLE',
}

export const MEDICAL_ROLES: Role[] = [
  Role.MEDICAL,
  Role.NURSE,
  Role.NURSING_TECHNICIAN,
  Role.PSYCHOLOGIST,
  Role.NUTRITIONIST,
]
export const SYSTEM_ROLES: Role[] = [Role.ADMIN, ...MEDICAL_ROLES]
export const MANAGER_ROLES: Role[] = [Role.COLABORATOR, Role.MANAGER]
export const STUDENT_ROLES: Role[] = [Role.STUDENT, Role.RESPONSABLE]
export const INSTITUTION_ROLES: Role[] = [...MANAGER_ROLES, ...STUDENT_ROLES]